Logics に処理を移動
