+++ app/backend/src/ScoreHistoryApi/Logics/IScoreItemStorage.cs
@@ -9,44 +9,5 @@ namespace ScoreHistoryApi.Logics
     /// </summary>
     public interface IScoreItemStorage
     {
-        /// <summary>
-        /// 楽譜のオブジェクトデータを保存する
-        /// </summary>
-        /// <param name="ownerId"></param>
-        /// <param name="scoreId"></param>
-        /// <param name="data"></param>
-        /// <param name="accessControl"></param>
-        /// <returns></returns>
-        Task<SavedItemData> SaveObjectAsync(Guid ownerId, Guid scoreId, byte[] data,
-            ScoreObjectAccessControls accessControl);
-
-        /// <summary>
-        /// 指定した楽譜のオブジェクトデータを削除する
-        /// </summary>
-        /// <param name="ownerId"></param>
-        /// <param name="scoreId"></param>
-        /// <param name="itemId"></param>
-        Task DeleteObjectAsync(Guid ownerId, Guid scoreId, Guid itemId);
-
-        /// <summary>
-        /// 指定した楽譜のオブジェクトデータを全て削除する
-        /// </summary>
-        /// <param name="ownerId"></param>
-        /// <param name="scoreId"></param>
-        Task DeleteAllScoreObjectAsync(Guid ownerId, Guid scoreId);
-
-        /// <summary>
-        /// 指定した owner のオブジェクトデータを全て削除する
-        /// </summary>
-        /// <param name="ownerId"></param>
-        Task DeleteAllOwnerObjectAsync(Guid ownerId);
-
-        /// <summary>
-        /// 指定した楽譜のオブジェクトのアクセスコントロールを設定する
-        /// </summary>
-        /// <param name="ownerId"></param>
-        /// <param name="scoreId"></param>
-        /// <param name="accessControl"></param>
-        Task SetAccessControlPolicyAsync(Guid ownerId, Guid scoreId, ScoreObjectAccessControls accessControl);
     }
 }
