+++ app/backend/src/ScoreHistoryApi/Logics/ScoreItemDatabase.cs
@@ -48,139 +48,6 @@ namespace ScoreHistoryApi.Logics
             _dynamoDbClient = dynamoDbClient;
         }
 
-        public async Task InitializeAsync(Guid ownerId)
-        {
-            var partitionKey = ScoreItemDatabaseUtils.ConvertToPartitionKey(ownerId);
-
-            await PutAsync(_dynamoDbClient, TableName, partitionKey);
-
-            static async Task PutAsync(IAmazonDynamoDB client, string tableName, string partitionKey)
-            {
-                var request = new PutItemRequest()
-                {
-                    Item = new Dictionary<string, AttributeValue>()
-                    {
-                        [ScoreItemDatabasePropertyNames.OwnerId] = new AttributeValue(partitionKey),
-                        [ScoreItemDatabasePropertyNames.ItemId] = new AttributeValue(ScoreItemDatabaseConstant.ItemIdSummary),
-                        [ScoreItemDatabasePropertyNames.Size] = new AttributeValue(){N = "0"},
-                    },
-                    TableName = tableName,
-                    ExpressionAttributeNames = new Dictionary<string, string>()
-                    {
-                        ["#owner"] = DynamoDbScorePropertyNames.OwnerId
-                    },
-                    ConditionExpression = "attribute_not_exists(#owner)",
-                };
-                try
-                {
-                    await client.PutItemAsync(request);
-                }
-                catch (ConditionalCheckFailedException ex)
-                {
-                    if (ex.ErrorCode == "ConditionalCheckFailedException")
-                    {
-                        throw new AlreadyInitializedException(ex);
-                    }
-                    throw;
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex.Message);
-                    throw;
-                }
-            }
-        }
-
-        public async Task CreateAsync(ScoreItemDatabaseItemDataBase itemData)
-        {
-
-            var now = ScoreDatabaseUtils.UnixTimeMillisecondsNow();
-
-            var maxSize = _quota.OwnerItemMaxSize;
-
-            await PutDataAsync(_dynamoDbClient, TableName, itemData, maxSize, now);
-
-            static async Task PutDataAsync(
-                IAmazonDynamoDB client,
-                string tableName,
-                ScoreItemDatabaseItemDataBase itemData,
-                long maxSize,
-                DateTimeOffset now)
-            {
-                var (items, partitionKey, _, item, totalSize) = ScoreItemDatabaseUtils.CreateDynamoDbValue(itemData, now);
-
-                var actions = new List<TransactWriteItem>()
-                {
-                    new TransactWriteItem()
-                    {
-                        Update = new Update()
-                        {
-                            Key = new Dictionary<string, AttributeValue>()
-                            {
-                                [ScoreItemDatabasePropertyNames.OwnerId] = new AttributeValue(partitionKey),
-                                [ScoreItemDatabasePropertyNames.ItemId] = new AttributeValue(ScoreItemDatabaseConstant.ItemIdSummary),
-                            },
-                            ExpressionAttributeNames = new Dictionary<string, string>()
-                            {
-                                ["#size"] = ScoreItemDatabasePropertyNames.Size,
-                            },
-                            ExpressionAttributeValues = new Dictionary<string, AttributeValue>()
-                            {
-                                [":itemSize"] = new AttributeValue(){N = totalSize.ToString()},
-                                [":maxSize"] = new AttributeValue()
-                                {
-                                    N = (maxSize - totalSize).ToString()
-                                },
-                            },
-                            ConditionExpression = "#size < :maxSize",
-                            UpdateExpression = "ADD #size :itemSize",
-                            TableName = tableName,
-                        },
-                    },
-                    new TransactWriteItem()
-                    {
-                        Put = new Put()
-                        {
-                            Item = items,
-                            TableName = tableName,
-                            ExpressionAttributeNames = new Dictionary<string, string>()
-                            {
-                                ["#item"] = ScoreItemDatabasePropertyNames.ItemId,
-                            },
-                            ConditionExpression = "attribute_not_exists(#item)",
-                        }
-                    },
-                };
-                try
-                {
-                    await client.TransactWriteItemsAsync(new TransactWriteItemsRequest()
-                    {
-                        TransactItems = actions,
-                        ReturnConsumedCapacity = ReturnConsumedCapacity.TOTAL
-                    });
-                }
-                catch (ResourceNotFoundException ex)
-                {
-                    Console.WriteLine(ex.Message);
-                    throw;
-                }
-                catch (InternalServerErrorException ex)
-                {
-                    Console.WriteLine(ex.Message);
-                    throw;
-                }
-                catch (TransactionCanceledException ex)
-                {
-                    Console.WriteLine(ex.Message);
-                    throw;
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex.Message);
-                    throw;
-                }
-            }
-        }
 
         public async Task DeleteAsync(Guid ownerId, Guid scoreId, Guid itemId)
         {
@@ -304,122 +171,6 @@ namespace ScoreHistoryApi.Logics
                     Console.WriteLine(ex.Message);
                     throw;
                 }
-            }
-        }
-
-        public async Task DeleteItemsAsync(Guid ownerId, DeletingScoreItems deletingScoreItems)
-        {
-            await CheckDeleteAsync(_dynamoDbClient, ScoreItemRelationTableName, ownerId, deletingScoreItems.ItemIds);
-
-            var itemAndSizeList = await GetItemAndSizeListAsync(_dynamoDbClient, TableName, ownerId, deletingScoreItems.ScoreId);
-
-            var targetHashSet = new HashSet<Guid>();
-            foreach (var itemId in deletingScoreItems.ItemIds)
-            {
-                targetHashSet.Add(itemId);
-            }
-
-            var filteredItemAndSizeList = itemAndSizeList
-                .Where(x => targetHashSet.Contains(x.itemId))
-                .Select(x => (x.itemIdText, x.size))
-                .ToArray();
-
-            await DeleteItemsAsync(ownerId, filteredItemAndSizeList);
-
-            static async Task<(string itemIdText, Guid itemId, long size)[]> GetItemAndSizeListAsync(IAmazonDynamoDB client, string tableName, Guid ownerId, Guid scoreId)
-            {
-                var partitionKey = ScoreItemDatabaseUtils.ConvertToPartitionKey(ownerId);
-                var score = ScoreDatabaseUtils.ConvertToBase64(scoreId);
-
-                var request = new QueryRequest()
-                {
-                    TableName = tableName,
-                    ExpressionAttributeNames = new Dictionary<string, string>()
-                    {
-                        ["#owner"] = ScoreItemDatabasePropertyNames.OwnerId,
-                        ["#item"] = ScoreItemDatabasePropertyNames.ItemId,
-                    },
-                    ExpressionAttributeValues = new Dictionary<string, AttributeValue>()
-                    {
-                        [":owner"] = new AttributeValue(partitionKey),
-                        [":score"] = new AttributeValue(score),
-                    },
-                    KeyConditionExpression = "#owner = :owner and begins_with(#item, :score)",
-                };
-                try
-                {
-                    var response = await client.QueryAsync(request);
-
-                    return response.Items
-                        .Where(x=>x[ScoreItemDatabasePropertyNames.ItemId].S != ScoreItemDatabaseConstant.ItemIdSummary)
-                        .Select(GetItemAndSizeAsync).ToArray();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex.Message);
-                    throw;
-                }
-
-                static (string itemIdText, Guid itemId, long size) GetItemAndSizeAsync(Dictionary<string, AttributeValue> items)
-                {
-                    var sizeValue = items[ScoreItemDatabasePropertyNames.Size];
-                    var size = long.Parse(sizeValue.N, CultureInfo.InvariantCulture);
-
-                    try
-                    {
-                        var type = items[ScoreItemDatabasePropertyNames.Type];
-                        if (type.S == ScoreItemDatabaseConstant.TypeImage)
-                        {
-                            var thumbnail = items[ScoreItemDatabasePropertyNames.Thumbnail];
-                            var thumbnailSizeValue =
-                                thumbnail.M[ScoreItemDatabasePropertyNames.ThumbnailPropertyNames.Size];
-                            size += long.Parse(thumbnailSizeValue.N, CultureInfo.InvariantCulture);
-                        }
-                    }
-                    catch (Exception ex)
-                    {
-                        throw;
-                    }
-
-
-                    var itemIdText = items[ScoreItemDatabasePropertyNames.ItemId].S;
-
-                    var item = itemIdText.Substring(ScoreItemDatabaseConstant.ScoreIdLength);
-                    var itemId = ScoreDatabaseUtils.ConvertToGuid(item);
-
-                    return (itemIdText, itemId, size);
-                }
-            }
-
-            static async Task CheckDeleteAsync(IAmazonDynamoDB client, string tableName, Guid ownerId, List<Guid> itemIds)
-            {
-                var partitionKey = DynamoDbScoreItemRelationUtils.ConvertToPartitionKey(ownerId);
-
-                foreach (var itemId in itemIds)
-                {
-                    var item = ScoreDatabaseUtils.ConvertToBase64(itemId);
-                    var request = new QueryRequest()
-                    {
-                        TableName = tableName,
-                        Limit = 1,
-                        ExpressionAttributeNames = new Dictionary<string, string>()
-                        {
-                            ["#owner"] = DynamoDbScoreItemRelationPropertyNames.OwnerId,
-                            ["#itemRel"] = DynamoDbScoreItemRelationPropertyNames.ItemRelation,
-                        },
-                        ExpressionAttributeValues = new Dictionary<string, AttributeValue>()
-                        {
-                            [":owner"] = new AttributeValue(partitionKey),
-                            [":item"] = new AttributeValue(item),
-                        },
-                        KeyConditionExpression = "#owner = :owner and begins_with(#itemRel, :item)",
-                    };
-                    var response = await client.QueryAsync(request);
-
-                    if (0 < response.Count)
-                        throw new InvalidOperationException($"'{itemId}' は参照されているため削除できません");
-                }
-
             }
         }
 
@@ -570,73 +321,6 @@ namespace ScoreHistoryApi.Logics
             }
         }
 
-        public async Task<ScoreItemDatabaseItemDataBase[]> GetItemsAsync(Guid ownerId)
-        {
-            return await GetAsync(_dynamoDbClient, TableName, ownerId);
-
-            static async Task<ScoreItemDatabaseItemDataBase[]> GetAsync(IAmazonDynamoDB client, string tableName, Guid ownerId)
-            {
-                var partitionKey = ScoreItemDatabaseUtils.ConvertToPartitionKey(ownerId);
-
-                var request = new QueryRequest()
-                {
-                    TableName = tableName,
-                    ExpressionAttributeNames = new Dictionary<string, string>()
-                    {
-                        ["#owner"] = ScoreItemDatabasePropertyNames.OwnerId,
-                    },
-                    ExpressionAttributeValues = new Dictionary<string, AttributeValue>()
-                    {
-                        [":owner"] = new AttributeValue(partitionKey),
-                    },
-                    KeyConditionExpression = "#owner = :owner",
-                    Limit = 500,
-                };
-
-                try
-                {
-                    var response = await client.QueryAsync(request);
-
-                    var items = response.Items.ToList();
-
-                    while(0 < response.LastEvaluatedKey?.Count)
-                    {
-                        var nextRequest = new QueryRequest()
-                        {
-                            TableName = tableName,
-                            ExpressionAttributeNames = new Dictionary<string, string>()
-                            {
-                                ["#owner"] = ScoreItemDatabasePropertyNames.OwnerId,
-                                ["#item"] = ScoreItemDatabasePropertyNames.ItemId,
-                            },
-                            ExpressionAttributeValues = new Dictionary<string, AttributeValue>()
-                            {
-                                [":owner"] = new AttributeValue(partitionKey),
-                                [":item"] = new AttributeValue(response
-                                    .LastEvaluatedKey[ScoreItemDatabasePropertyNames.ItemId].S),
-                            },
-                            KeyConditionExpression = "#owner = :owner and :item < #item",
-                            Limit = 500,
-                        };
-
-                        var nextResponse = await client.QueryAsync(nextRequest);
-                        items.AddRange(nextResponse.Items);
-
-                        response = nextResponse;
-                    }
-
-                    return items
-                        .Where(x=>x[ScoreItemDatabasePropertyNames.ItemId].S != ScoreItemDatabaseConstant.ItemIdSummary)
-                        .Select(ScoreItemDatabaseUtils.ConvertFromDynamoDbValue)
-                        .ToArray();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex.Message);
-                    throw;
-                }
-            }
-        }
 
 
         public async Task DeleteItemsAsync(Guid ownerId, (string itemIdText, long size)[] itemAndSizeList)
