+++ app/backend/src/ScoreHistoryApi/Logics/ScoreItemStorage.cs
@@ -122,88 +122,6 @@ namespace ScoreHistoryApi.Logics
             _s3Client = s3Client;
         }
 
-        public async Task<SavedItemData> SaveObjectAsync(
-            Guid ownerId, Guid scoreId, byte[] data,
-            ScoreObjectAccessControls accessControl)
-        {
-            return await SaveObjectAsync(ownerId, scoreId, Guid.NewGuid(), data, accessControl);
-        }
-        public async Task<SavedItemData> SaveObjectAsync(Guid ownerId, Guid scoreId, Guid itemId, byte[] data, ScoreObjectAccessControls accessControl)
-        {
-            if (data is null)
-            {
-                throw new ArgumentNullException(nameof(data));
-            }
-            var itemType = ScoreItemStorageUtils.CheckItemType(data);
-
-            if (itemType == ItemTypes.None)
-                throw new NotSupportedItemFileException($"Not supported file. top 10 bytes : '{string.Join(" ",data.Take(10).Select(b=>b.ToString("X2")))}'");
-
-            var keyDir = $"{ownerId:D}/{scoreId:D}/{ScoreItemStorageConstant.FolderName}/{itemId:D}/";
-
-
-            await using var srcStream = new MemoryStream(data);
-
-            using var thumbnailImage = await Image.LoadAsync(srcStream);
-            srcStream.Seek(0, SeekOrigin.Begin);
-
-
-            var height = (int) (thumbnailImage.Height * (double) ScoreItemStorageConstant.ThumbnailMaxWidthPixel / thumbnailImage.Width);
-            thumbnailImage.Mutate(x=>x.Resize(ScoreItemStorageConstant.ThumbnailMaxWidthPixel,height));
-
-            await using var thumbnailStream = new MemoryStream();
-
-            await thumbnailImage.SaveAsJpegAsync(thumbnailStream);
-            thumbnailStream.Seek(0, SeekOrigin.Begin);
-
-            var thumbnailSize = thumbnailStream.Length;
-
-            var objectFileName = ScoreItemStorageUtils.GetFileName(itemType);
-            var thumbnailFileName = ScoreItemStorageConstant.ThumbnailFileName;
-
-            var acl = accessControl switch
-            {
-                ScoreObjectAccessControls.Private => S3CannedACL.Private,
-                ScoreObjectAccessControls.Public => S3CannedACL.PublicRead,
-                _ => throw new NotSupportedException(),
-            };
-
-            var objectSaveRequest = new PutObjectRequest()
-            {
-                BucketName = BucketName,
-                Key = keyDir + objectFileName,
-                CannedACL = acl,
-                InputStream = srcStream,
-            };
-
-            var thumbnailSaveRequest = new PutObjectRequest()
-            {
-                BucketName = BucketName,
-                Key = keyDir + thumbnailFileName,
-                CannedACL = acl,
-                InputStream = thumbnailStream,
-            };
-
-            await _s3Client.PutObjectAsync(thumbnailSaveRequest);
-            await _s3Client.PutObjectAsync(objectSaveRequest);
-
-            return new SavedItemData()
-            {
-                Data = data,
-                OwnerId = ownerId,
-                ScoreId = scoreId,
-                Type = itemType,
-                ItemId = itemId,
-                ObjectName = objectFileName,
-                Size = data.Length,
-                Extra = new Thumbnail()
-                {
-                    ObjectName = thumbnailFileName,
-                    Size = thumbnailSize,
-                },
-                AccessControl = ScoreObjectAccessControls.Private
-            };
-        }
 
 
         public async Task DeleteObjectAsync(Guid ownerId, Guid scoreId, Guid itemId)
@@ -256,49 +174,5 @@ namespace ScoreHistoryApi.Logics
             await _s3Client.DeleteObjectsAsync(request);
         }
 
-        public async Task SetAccessControlPolicyAsync(Guid ownerId, Guid scoreId,
-            ScoreObjectAccessControls accessControl)
-        {
-            var prefix = $"{ownerId:D}/{scoreId:D}/{ScoreItemStorageConstant.FolderName}";
-
-            var objectKeyList = new List<string>();
-            string continuationToken = default;
-
-            do
-            {
-                var listRequest = new ListObjectsV2Request()
-                {
-                    BucketName = BucketName,
-                    Prefix = prefix,
-                    ContinuationToken = string.IsNullOrWhiteSpace(continuationToken) ? null : continuationToken,
-                };
-                var listResponse = await _s3Client.ListObjectsV2Async(listRequest);
-
-                objectKeyList.AddRange(listResponse.S3Objects.Select(x => x.Key));
-
-                continuationToken = listResponse.NextContinuationToken;
-
-            } while (!string.IsNullOrEmpty(continuationToken));
-
-
-            var acl = accessControl switch
-            {
-                ScoreObjectAccessControls.Private => S3CannedACL.Private,
-                ScoreObjectAccessControls.Public => S3CannedACL.PublicRead,
-                _ => throw new NotSupportedException(),
-            };
-
-            foreach (var key in objectKeyList)
-            {
-                var request = new PutACLRequest()
-                {
-                    BucketName = BucketName,
-                    CannedACL = acl,
-                    Key = key,
-                };
-                await _s3Client.PutACLAsync(request);
-            }
-
-        }
     }
 }
