+++ app/backend/src/ScoreHistoryApi/Logics/ScoreItems/ScoreItemInfoGetter.cs
@@ -1,6 +1,10 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Amazon.DynamoDBv2;
+using Amazon.DynamoDBv2.Model;
+using Microsoft.Extensions.Configuration;
 using ScoreHistoryApi.Logics.ScoreItemDatabases;
 using ScoreHistoryApi.Models.ScoreItems;
 
@@ -8,16 +12,24 @@ namespace ScoreHistoryApi.Logics.ScoreItems
 {
     public class ScoreItemInfoGetter
     {
-        private readonly IScoreItemDatabase _scoreItemDatabase;
+        private readonly IAmazonDynamoDB _dynamoDbClient;
 
-        public ScoreItemInfoGetter(IScoreItemDatabase scoreItemDatabase)
+        public ScoreItemInfoGetter(IAmazonDynamoDB dynamoDbClient, IConfiguration configuration)
         {
-            _scoreItemDatabase = scoreItemDatabase;
+            _dynamoDbClient = dynamoDbClient;
+
+            var tableName = configuration[EnvironmentNames.ScoreItemDynamoDbTableName];
+            if (string.IsNullOrWhiteSpace(tableName))
+                throw new InvalidOperationException($"'{EnvironmentNames.ScoreItemDynamoDbTableName}' is not found.");
+
+            TableName = tableName;
         }
+
+        public string TableName { get; set; }
 
         public async Task<OwnerItemsInfo> GetOwnerItemsInfoAsync(Guid ownerId)
         {
-            var itemDataList = await _scoreItemDatabase.GetItemsAsync(ownerId);
+            var itemDataList = await GetItemsAsync(ownerId);
 
             var totalSize = itemDataList.Sum(x => x.TotalSize);
 
@@ -32,7 +44,7 @@ namespace ScoreHistoryApi.Logics.ScoreItems
 
         public async Task<UserItemsInfo> GetUserItemsInfoAsync(Guid ownerId)
         {
-            var itemDataList = await _scoreItemDatabase.GetItemsAsync(ownerId);
+            var itemDataList = await GetItemsAsync(ownerId);
 
             var totalSize = itemDataList.Sum(x => x.TotalSize);
 
@@ -63,6 +75,75 @@ namespace ScoreHistoryApi.Logics.ScoreItems
             }
 
             throw new ArgumentException();
+        }
+
+
+        public async Task<ScoreItemDatabaseItemDataBase[]> GetItemsAsync(Guid ownerId)
+        {
+            return await GetAsync(_dynamoDbClient, TableName, ownerId);
+
+            static async Task<ScoreItemDatabaseItemDataBase[]> GetAsync(IAmazonDynamoDB client, string tableName, Guid ownerId)
+            {
+                var partitionKey = ScoreItemDatabaseUtils.ConvertToPartitionKey(ownerId);
+
+                var request = new QueryRequest()
+                {
+                    TableName = tableName,
+                    ExpressionAttributeNames = new Dictionary<string, string>()
+                    {
+                        ["#owner"] = ScoreItemDatabasePropertyNames.OwnerId,
+                    },
+                    ExpressionAttributeValues = new Dictionary<string, AttributeValue>()
+                    {
+                        [":owner"] = new AttributeValue(partitionKey),
+                    },
+                    KeyConditionExpression = "#owner = :owner",
+                    Limit = 500,
+                };
+
+                try
+                {
+                    var response = await client.QueryAsync(request);
+
+                    var items = response.Items.ToList();
+
+                    while(0 < response.LastEvaluatedKey?.Count)
+                    {
+                        var nextRequest = new QueryRequest()
+                        {
+                            TableName = tableName,
+                            ExpressionAttributeNames = new Dictionary<string, string>()
+                            {
+                                ["#owner"] = ScoreItemDatabasePropertyNames.OwnerId,
+                                ["#item"] = ScoreItemDatabasePropertyNames.ItemId,
+                            },
+                            ExpressionAttributeValues = new Dictionary<string, AttributeValue>()
+                            {
+                                [":owner"] = new AttributeValue(partitionKey),
+                                [":item"] = new AttributeValue(response
+                                    .LastEvaluatedKey[ScoreItemDatabasePropertyNames.ItemId].S),
+                            },
+                            KeyConditionExpression = "#owner = :owner and :item < #item",
+                            Limit = 500,
+                        };
+
+                        var nextResponse = await client.QueryAsync(nextRequest);
+                        items.AddRange(nextResponse.Items);
+
+                        response = nextResponse;
+                    }
+
+                    return items
+                        .Where(x=>x[ScoreItemDatabasePropertyNames.ItemId].S != ScoreItemDatabaseConstant.ItemIdSummary)
+                        .Select(ScoreItemDatabaseUtils.ConvertFromDynamoDbValue)
+                        .ToArray();
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                    throw;
+                }
+            }
         }
     }
 }
