+++ app/backend/src/ScoreHistoryApi/Logics/ScoreSnapshotStorage.cs
@@ -86,31 +86,6 @@ namespace ScoreHistoryApi.Logics
             _s3Client = s3Client;
         }
 
-        public async Task CreateAsync(Guid ownerId, Guid scoreId, ScoreSnapshotDetail snapshotDetail,
-            ScoreObjectAccessControls accessControl)
-        {
-            var key = ScoreSnapshotStorageUtils.CreateSnapshotKey(ownerId, scoreId, snapshotDetail.Id);
-
-            var json = ScoreSnapshotStorageUtils.ConvertToJson(snapshotDetail);
-
-            await using var jsonStream = new MemoryStream(json);
-
-            var acl = accessControl switch
-            {
-                ScoreObjectAccessControls.Private => S3CannedACL.Private,
-                ScoreObjectAccessControls.Public => S3CannedACL.PublicRead,
-                _ => throw new NotSupportedException(),
-            };
-
-            var request = new PutObjectRequest()
-            {
-                BucketName = BucketName,
-                Key = key,
-                CannedACL = acl,
-                InputStream = jsonStream,
-            };
-            await _s3Client.PutObjectAsync(request);
-        }
 
         public async Task DeleteAsync(Guid ownerId, Guid scoreId, Guid snapshotId)
         {
@@ -124,109 +99,5 @@ namespace ScoreHistoryApi.Logics
             var response = await _s3Client.DeleteObjectAsync(request);
         }
 
-        public async Task DeleteAllAsync(Guid ownerId, Guid scoreId)
-        {
-            var objectKeyList = new List<string>();
-            string continuationToken = default;
-
-            var prefix = $"{ownerId:D}/{scoreId:D}/{ScoreSnapshotStorageConstant.SnapshotFolderName}";
-
-            do
-            {
-                var listRequest = new ListObjectsV2Request()
-                {
-                    BucketName = BucketName,
-                    Prefix = prefix,
-                    ContinuationToken = string.IsNullOrWhiteSpace(continuationToken) ? null : continuationToken,
-                };
-                var listResponse = await _s3Client.ListObjectsV2Async(listRequest);
-
-                objectKeyList.AddRange(listResponse.S3Objects.Select(x => x.Key));
-
-                continuationToken = listResponse.NextContinuationToken;
-
-            } while (!string.IsNullOrEmpty(continuationToken));
-
-            var request = new DeleteObjectsRequest()
-            {
-                BucketName = BucketName,
-                Objects = objectKeyList.Select(x=>new KeyVersion()
-                {
-                    Key = x
-                }).ToList(),
-            };
-            await _s3Client.DeleteObjectsAsync(request);
-        }
-
-        public async Task SetAccessControlPolicyAsync(Guid ownerId, Guid scoreId, ScoreObjectAccessControls accessControl)
-        {
-            var prefix = $"{ownerId:D}/{scoreId:D}/{ScoreSnapshotStorageConstant.SnapshotFolderName}";
-
-            var objectKeyList = new List<string>();
-            string continuationToken = default;
-
-            do
-            {
-                var listRequest = new ListObjectsV2Request()
-                {
-                    BucketName = BucketName,
-                    Prefix = prefix,
-                    ContinuationToken = string.IsNullOrWhiteSpace(continuationToken) ? null : continuationToken,
-                };
-                var listResponse = await _s3Client.ListObjectsV2Async(listRequest);
-
-                objectKeyList.AddRange(listResponse.S3Objects.Select(x => x.Key));
-
-                continuationToken = listResponse.NextContinuationToken;
-
-            } while (!string.IsNullOrEmpty(continuationToken));
-
-
-            var acl = accessControl switch
-            {
-                ScoreObjectAccessControls.Private => S3CannedACL.Private,
-                ScoreObjectAccessControls.Public => S3CannedACL.PublicRead,
-                _ => throw new NotSupportedException(),
-            };
-
-            foreach (var key in objectKeyList)
-            {
-                var request = new PutACLRequest()
-                {
-                    BucketName = BucketName,
-                    CannedACL = acl,
-                    Key = key,
-                };
-                await _s3Client.PutACLAsync(request);
-            }
-        }
-
-        public async Task<ScoreSnapshotDetail> GetAsync(Guid ownerId, Guid scoreId, Guid snapshotId)
-        {
-            var key = ScoreSnapshotStorageUtils.CreateSnapshotKey(ownerId, scoreId, snapshotId);
-
-            var request = new GetObjectRequest()
-            {
-                BucketName = BucketName,
-                Key = key,
-            };
-            try
-            {
-                var response = await _s3Client.GetObjectAsync(request);
-
-                if (response.HttpStatusCode == HttpStatusCode.OK)
-                {
-                    byte[] buffer = new byte[response.ResponseStream.Length];
-                    await response.ResponseStream.ReadAsync(buffer);
-
-                    return ScoreSnapshotStorageUtils.MapFromJson(buffer);
-                }
-            }
-            catch (Exception ex)
-            {
-                throw new NotFoundSnapshotException(ex);
-            }
-            throw new NotFoundSnapshotException("Not found snapshot.");
-        }
     }
 }
