+++ app/backend/src/ScoreHistoryApi/Logics/Scores/ScorePageAdder.cs
@@ -1,18 +1,43 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
+using Amazon.DynamoDBv2;
+using Amazon.DynamoDBv2.Model;
+using Microsoft.Extensions.Configuration;
+using ScoreHistoryApi.Logics.ScoreDatabases;
 using ScoreHistoryApi.Models.Scores;
 
 namespace ScoreHistoryApi.Logics.Scores
 {
     public class ScorePageAdder
     {
-        private readonly IScoreDatabase _scoreDatabase;
+        private readonly IAmazonDynamoDB _dynamoDbClient;
+        private readonly IScoreQuota _scoreQuota;
+        private readonly IConfiguration _configuration;
 
-        public ScorePageAdder(IScoreDatabase scoreDatabase)
+        public ScorePageAdder(IAmazonDynamoDB dynamoDbClient, IScoreQuota scoreQuota, IConfiguration configuration)
         {
-            _scoreDatabase = scoreDatabase;
+            _dynamoDbClient = dynamoDbClient;
+            _scoreQuota = scoreQuota;
+            _configuration = configuration;
+
+
+            var tableName = configuration[EnvironmentNames.ScoreDynamoDbTableName];
+            if (string.IsNullOrWhiteSpace(tableName))
+                throw new InvalidOperationException($"'{EnvironmentNames.ScoreDynamoDbTableName}' is not found.");
+            ScoreTableName = tableName;
+
+            var scoreItemRelationTableName = configuration[EnvironmentNames.ScoreItemRelationDynamoDbTableName];
+            if (string.IsNullOrWhiteSpace(scoreItemRelationTableName))
+                throw new InvalidOperationException(
+                    $"'{EnvironmentNames.ScoreItemRelationDynamoDbTableName}' is not found.");
+            ScoreItemRelationTableName = scoreItemRelationTableName;
         }
+
+        public string ScoreItemRelationTableName { get; set; }
+
+        public string ScoreTableName { get; set; }
 
         public async Task AddPages(Guid ownerId, Guid scoreId, List<NewScorePage> pages)
         {
@@ -46,7 +71,214 @@ namespace ScoreHistoryApi.Logics.Scores
                 });
             }
 
-            await _scoreDatabase.AddPagesAsync(ownerId, scoreId, trimmedPages);
+            await AddPagesAsync(ownerId, scoreId, trimmedPages);
+        }
+
+
+        public async Task AddPagesAsync(Guid ownerId, Guid scoreId, List<NewScorePage> pages)
+        {
+            if (pages.Count == 0)
+                throw new ArgumentException(nameof(pages));
+
+            var (data, oldHash) = await GetAsync(_dynamoDbClient, ScoreTableName, ownerId, scoreId);
+
+            data.Page ??= new List<DynamoDbScorePageV1>();
+
+            var newPages = new List<DynamoDbScorePageV1>();
+            var newItemRelationSet = new HashSet<string>();
+
+            var pageId = data.Page.Count == 0 ? 0 : data.Page.Select(x => x.Id).Max() + 1;
+            foreach (var page in pages)
+            {
+                var itemId = ScoreDatabaseUtils.ConvertToBase64(page.ItemId);
+                var p = new DynamoDbScorePageV1()
+                {
+                    Id = pageId++,
+                    ItemId = itemId,
+                    Page = page.Page,
+                    ObjectName = page.ObjectName,
+                };
+                newPages.Add(p);
+                data.Page.Add(p);
+                newItemRelationSet.Add(itemId);
+            }
+
+            var newHash = data.CalcDataHash();
+
+            var now = ScoreDatabaseUtils.UnixTimeMillisecondsNow();
+
+            // TODO ページの追加上限値判定を追加
+            await UpdateAsync(_dynamoDbClient, ScoreTableName, ownerId, scoreId, newPages, newHash, oldHash, now);
+
+            await PutItemRelations(_dynamoDbClient, ScoreItemRelationTableName, ownerId, scoreId, newItemRelationSet);
+
+
+            static async Task<(DynamoDbScoreDataV1 data,string hash)> GetAsync(
+                IAmazonDynamoDB client,
+                string tableName,
+                Guid ownerId,
+                Guid scoreId)
+            {
+                var partitionKey = ScoreDatabaseUtils.ConvertToPartitionKey(ownerId);
+                var score = ScoreDatabaseUtils.ConvertToBase64(scoreId);
+
+                var request = new GetItemRequest()
+                {
+                    TableName = tableName,
+                    Key = new Dictionary<string, AttributeValue>()
+                    {
+                        [DynamoDbScorePropertyNames.OwnerId] = new AttributeValue(partitionKey),
+                        [DynamoDbScorePropertyNames.ScoreId] = new AttributeValue(ScoreDatabaseConstant.ScoreIdMainPrefix + score),
+                    },
+                };
+                var response = await client.GetItemAsync(request);
+                var data = response.Item[DynamoDbScorePropertyNames.Data];
+
+                if (data is null)
+                    throw new InvalidOperationException("not found.");
+
+
+                DynamoDbScoreDataV1.TryMapFromAttributeValue(data, out var result);
+                var hash = response.Item[DynamoDbScorePropertyNames.DataHash].S;
+                return (result,hash);
+            }
+
+            static AttributeValue ConvertFromPages(List<DynamoDbScorePageV1> pages)
+            {
+                var result = new AttributeValue()
+                {
+                    L = new List<AttributeValue>()
+                };
+
+                foreach (var page in pages)
+                {
+                    var p = new Dictionary<string, AttributeValue>()
+                    {
+                        [DynamoDbScorePropertyNames.DataPropertyNames.PagesPropertyNames.Id] = new AttributeValue()
+                        {
+                            N = page.Id.ToString(),
+                        }
+                    };
+                    if (page.Page != null)
+                    {
+                        p[DynamoDbScorePropertyNames.DataPropertyNames.PagesPropertyNames.Page] = new AttributeValue(page.Page);
+                    }
+                    if (page.ItemId != null)
+                    {
+                        p[DynamoDbScorePropertyNames.DataPropertyNames.PagesPropertyNames.ItemId] = new AttributeValue(page.ItemId);
+                    }
+                    if (page.ObjectName != null)
+                    {
+                        p[DynamoDbScorePropertyNames.DataPropertyNames.PagesPropertyNames.ObjectName] = new AttributeValue(page.ObjectName);
+                    }
+                    if(p.Count == 0)
+                        continue;
+
+                    result.L.Add(new AttributeValue() {M = p});
+                }
+
+                return result;
+            }
+
+            static async Task UpdateAsync(
+                IAmazonDynamoDB client,
+                string tableName,
+                Guid ownerId,
+                Guid scoreId,
+                List<DynamoDbScorePageV1> newPages,
+                string newHash,
+                string oldHash,
+                DateTimeOffset now
+                )
+            {
+                var partitionKey = ScoreDatabaseUtils.ConvertToPartitionKey(ownerId);
+                var score = ScoreDatabaseUtils.ConvertToBase64(scoreId);
+
+                var updateAt = ScoreDatabaseUtils.ConvertToUnixTimeMilli(now);
+
+                var request = new UpdateItemRequest()
+                {
+                    Key = new Dictionary<string, AttributeValue>()
+                    {
+                        [DynamoDbScorePropertyNames.OwnerId] = new AttributeValue(partitionKey),
+                        [DynamoDbScorePropertyNames.ScoreId] = new AttributeValue(ScoreDatabaseConstant.ScoreIdMainPrefix + score),
+                    },
+                    ExpressionAttributeNames = new Dictionary<string, string>()
+                    {
+                        ["#updateAt"] = DynamoDbScorePropertyNames.UpdateAt,
+                        ["#hash"] = DynamoDbScorePropertyNames.DataHash,
+                        ["#data"] = DynamoDbScorePropertyNames.Data,
+                        ["#pages"] = DynamoDbScorePropertyNames.DataPropertyNames.Pages,
+                    },
+                    ExpressionAttributeValues = new Dictionary<string, AttributeValue>()
+                    {
+                        [":newPages"] = ConvertFromPages(newPages),
+                        [":newHash"] = new AttributeValue(newHash),
+                        [":oldHash"] = new AttributeValue(oldHash),
+                        [":updateAt"] = new AttributeValue(updateAt),
+                    },
+                    ConditionExpression = "#hash = :oldHash",
+                    UpdateExpression = "SET #updateAt = :updateAt, #hash = :newHash, #data.#pages = list_append(#data.#pages, :newPages)",
+                    TableName = tableName,
+                };
+                try
+                {
+                    await client.UpdateItemAsync(request);
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                    throw;
+                }
+            }
+
+            static async Task PutItemRelations(IAmazonDynamoDB client, string tableName, Guid ownerId, Guid scoreId, HashSet<string> items)
+            {
+                const int chunkSize = 25;
+
+                var partitionKey = DynamoDbScoreItemRelationUtils.ConvertToPartitionKey(ownerId);
+                var score = ScoreDatabaseUtils.ConvertToBase64(scoreId);
+
+                var chunkList = items.Select((x, index) => (x, index))
+                    .GroupBy(x => x.index / chunkSize)
+                    .Select(x=>x.Select(y=>y.x).ToArray())
+                    .ToArray();
+
+                foreach (var ids in chunkList)
+                {
+                    await PutItemRelations25Async(client, tableName, partitionKey, score, ids);
+                }
+
+                static async Task PutItemRelations25Async(IAmazonDynamoDB client, string tableName, string partitionKey, string score, string[] ids)
+                {
+                    var request = new Dictionary<string, List<WriteRequest>>()
+                    {
+                        [tableName] = ids.Select(id=>new WriteRequest()
+                        {
+                            PutRequest = new PutRequest()
+                            {
+                                Item = new Dictionary<string, AttributeValue>()
+                                {
+                                    [DynamoDbScoreItemRelationPropertyNames.OwnerId] = new AttributeValue(partitionKey),
+                                    [DynamoDbScoreItemRelationPropertyNames.ItemRelation] = new AttributeValue(id + score),
+                                }
+                            }
+
+                        }).ToList(),
+                    };
+
+                    try
+                    {
+                        await client.BatchWriteItemAsync(request);
+                    }
+                    catch (Exception ex)
+                    {
+                        // TODO 削除時に失敗したデータを取得しリトライ処理を入れる
+                        Console.WriteLine(ex.Message);
+                        throw;
+                    }
+                }
+            }
         }
     }
 }
